began major refactor clean up
