+++ render.py
@@ -1,20 +1,13 @@
-import gc, uuid, math
+import uuid
 import queue
 import ffmpeg
 import PIL.Image
 import numpy as np
 import torch as th
-import torch.utils.data as data
 from threading import Thread
 
 th.set_grad_enabled(False)
 th.backends.cudnn.benchmark = True
-
-
-class Print(th.nn.Module):
-    def forward(self, x, *args, **kwargs):
-        print(x.shape)
-        return x
 
 
 def render(
@@ -25,16 +18,17 @@ def render(
     duration,
     batch_size,
     out_size,
+    output_file,
     audio_file=None,
     truncation=1,
-    manipulations=[],
-    output_file=None,
+    bends=[],
+    rewrites={},
+    randomize_noise=False,
 ):
-    output_file = output_file if output_file is not None else f"/home/hans/neurout/{uuid.uuid4().hex[:8]}.mp4"
-
     split_queue = queue.Queue()
     render_queue = queue.Queue()
 
+    # postprocesses batched torch tensors to individual RGB numpy arrays
     def split_batches(jobs_in, jobs_out):
         while True:
             try:
@@ -47,11 +41,12 @@ def render(
                 jobs_out.put(img.cpu().numpy().astype(np.uint8))
             jobs_in.task_done()
 
-    res = "1024x1024" if out_size == 1024 else ("512x512" if out_size == 512 else "1920x1080")
+    # start background ffmpeg process that listens on stdin for frame data
+    output_size = "1024x1024" if out_size == 1024 else ("512x512" if out_size == 512 else "1920x1080")
     if audio_file is not None:
         audio = ffmpeg.input(audio_file, ss=offset, to=offset + duration, guess_layout_max=0)
         video = (
-            ffmpeg.input("pipe:", format="rawvideo", pix_fmt="rgb24", framerate=len(latents) / duration, s=res)
+            ffmpeg.input("pipe:", format="rawvideo", pix_fmt="rgb24", framerate=len(latents) / duration, s=output_size)
             .output(
                 audio,
                 output_file,
@@ -62,19 +57,20 @@ def render(
                 ac=2,
                 v="warning",
             )
-            .global_args("-benchmark", "-stats", "-hide_banner")
+            .global_args("-hide_banner")
             .overwrite_output()
             .run_async(pipe_stdin=True)
         )
     else:
         video = (
-            ffmpeg.input("pipe:", format="rawvideo", pix_fmt="rgb24", framerate=len(latents) / duration, s=res)
+            ffmpeg.input("pipe:", format="rawvideo", pix_fmt="rgb24", framerate=len(latents) / duration, s=output_size)
             .output(output_file, framerate=len(latents) / duration, vcodec="libx264", preset="slow", v="warning",)
-            .global_args("-benchmark", "-stats", "-hide_banner")
+            .global_args("-hide_banner")
             .overwrite_output()
             .run_async(pipe_stdin=True)
         )
 
+    # writes numpy frames to ffmpeg stdin as raw rgb24 bytes
     def make_video(jobs_in):
         for _ in range(len(latents)):
             img = jobs_in.get(timeout=10)
@@ -82,6 +78,10 @@ def render(
                 img = img[:, 112:-112, :]
                 im = PIL.Image.fromarray(img)
                 img = np.array(im.resize((1920, 1080), PIL.Image.BILINEAR))
+            assert (
+                img.shape[1] == int(output_size.split("x")[1]) and img.shape[2] == int(output_size.split("x")[0]),
+                "generator's output image size does not match specified output size",
+            )
             video.stdin.write(img.tobytes())
             jobs_in.task_done()
         video.stdin.close()
@@ -92,85 +92,62 @@ def render(
     renderer = Thread(target=make_video, args=(render_queue,))
     renderer.daemon = True
 
+    # make all data that needs to be loaded to the GPU float, contiguous, and pinned
+    # the entire process is severly memory-transfer bound, but at least this might help a little
     latents = latents.float().contiguous().pin_memory()
-
-    rewrite_env = (
-        th.cat(
-            [
-                th.zeros((int(len(latents) / 12))),
-                th.linspace(0, 1, int(len(latents) / 12)) ** 1.75,
-                th.linspace(1, 0, int(len(latents) / 12)) ** 3,
-                th.linspace(0, 0.3, int(len(latents) / 24)),
-                th.linspace(0.3, 1, int(len(latents) / 48)),
-                th.linspace(1, 0, int(len(latents) / 48)),
-                th.zeros((int(3 * len(latents) / 24))),
-                th.linspace(0, 1, int(len(latents) / 48)),
-                th.linspace(1, 0, int(len(latents) / 48)),
-                th.zeros((int(len(latents) / 12))),
-                1 - th.linspace(1, 0, int(len(latents) / 12)) ** 2,
-                th.linspace(1, 0, int(len(latents) / 3)),
-            ],
-            axis=0,
-        )
-        .float()
-        .contiguous()
-        .pin_memory()
-    )
-    rewrite_env = th.cat([rewrite_env, th.zeros((len(latents) - len(rewrite_env)))]) ** 1.5
-
-    orig_weights = [getattr(generator.convs, f"{i}").conv.weight.clone() for i in range(len(generator.convs)) if i <= 7]
-    [print(ogw.shape) for ogw in orig_weights]
-    _, filin, filout, kh, kw = orig_weights[0].shape
-
-    from audiovisual import gaussian_filter
-
-    rewrite_noise = gaussian_filter(th.randn((len(latents), filin * filout, kh, kw)) - 1, 3)
-    print(rewrite_noise.min(), rewrite_noise.mean(), rewrite_noise.max())
-    rewrite_noise = rewrite_noise.reshape((len(latents), filin, filout, kh, kw)).float().contiguous().pin_memory()
 
     for ni, noise_scale in enumerate(noise):
         noise[ni] = noise_scale.float().contiguous().pin_memory() if noise_scale is not None else None
 
+    param_dict = dict(generator.named_parameters())
+    for param, (transform, modulation) in rewrites.items():
+        rewrites[param] = [transform, modulation.float().contiguous().pin_memory()]
+        original_weights[param] = param_dict[param].copy().cpu().float().contiguous().pin_memory()
+
+    for bend in bends:
+        if "modulation" in bend:
+            bend["modulation"] = bend["modulation"].float().contiguous().pin_memory()
+
+    if not isinstance(truncation, float):
+        truncation = truncation.float().contiguous().pin_memory()
+
     for n in range(0, len(latents), batch_size):
+        # load batches of data onto the GPU
         latent_batch = latents[n : n + batch_size].cuda(non_blocking=True)
-        noise_batch = [
-            (noise_scale[n : n + batch_size].cuda(non_blocking=True) if noise_scale is not None else None)
-            for noise_scale in noise
-        ]
 
-        manipulation_batch = []
-        if manipulations is not None:
-            for manip in manipulations:
-                if "params" in manip:
-                    manipulation_batch.append(
-                        {
-                            "layer": manip["layer"],
-                            "transform": manip["transform"](
-                                manip["params"][n : n + batch_size].cuda(non_blocking=True)
-                            ),
-                        }
-                    )
+        noise_batch = []
+        for noise_scale in noise:
+            if noise_scale is not None:
+                noise_batch.append(noise_scale[n : n + batch_size].cuda(non_blocking=True))
+            else:
+                noise_batch.append(None)
+
+        bend_batch = []
+        if bends is not None:
+            for bend in bends:
+                if "modulation" in bend:
+                    transform = bend["transform"](bend["modulation"][n : n + batch_size].cuda(non_blocking=True))
+                    bend_batch.append({"layer": bend["layer"], "transform": transform})
                 else:
-                    manipulation_batch.append({"layer": manip["layer"], "transform": manip["transform"]})
+                    bend_batch.append({"layer": bend["layer"], "transform": bend["transform"]})
 
-        for i in range(len(generator.convs)):
-            if i <= 7:
-                rewrite_env_batch = rewrite_env[n : n + batch_size, None, None, None, None].cuda(non_blocking=True)
-                rewrite_noise_batch = rewrite_noise[n : n + batch_size].cuda(non_blocking=True)
-                rewritten_weight = (1 - rewrite_env_batch) * orig_weights[i] + 2 * rewrite_env_batch * orig_weights[
-                    i
-                ] * rewrite_noise_batch
-                setattr(getattr(generator.convs, f"{i}").conv, "weight", th.nn.Parameter(rewritten_weight))
+        for param, rewrite in rewrites.items():
+            rewritten_weight = rewrite(original_weights[param], n,).cuda(non_blocking=True)
+            setattr(generator, param, th.nn.Parameter(rewritten_weight))
 
+        truncation_batch = truncation[n : n + batch_size] if not isinstance(truncation, float) else truncation
+
+        # forward through the generator
         outputs, _ = generator(
             styles=latent_batch,
             noise=noise_batch,
-            truncation=truncation,
-            transform_dict_list=manipulation_batch,
-            randomize_noise=False,
+            truncation=truncation_batch,
+            transform_dict_list=bend_batch,
+            randomize_noise=randomize_noise,
             input_is_latent=True,
         )
 
+        # send output to be split into frames and rendered one by one
         split_queue.put(outputs)
 
         if n == 0:
